DYNAMIC DATA DISPLAY
The data on both home page and on the reviews on a single game are fetched and set. Error on the display of the reviews needs a fix. Pictures are not loaded yet as at the moment static picture is used until further progress with the finishing fully the display on a nice view.