ADD GAME CAM OR GALLERY
Adding games to the API with the option of either a camera or a picture from the user gallery. Error handling in case the user forgets to add a field or leaves it empty. Refactoring and changing the file of Single Game view to work with scroll view. Markdown fields added for both add new game and review section.